dbpush and get

+++ src/main/java/com/project/repository/InvoiceRepository.java
@@ -3,5 +3,5 @@ package com.project.repository;
 import com.project.entity.Invoice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface InvoiceRepository extends JpaRepository<Invoice, Integer> {
+public interface InvoiceRepository extends JpaRepository<Invoice, Long> {
 }
